turned k8s reconciler into operator resource
