+++ service/resource/legacy/error.go
@@ -1,8 +1,15 @@
-package k8s
+package legacy
 
 import (
 	"github.com/juju/errgo"
 )
+
+var executionFailedError = errgo.New("execution failed")
+
+// IsExecutionFailed asserts executionFailedError.
+func IsExecutionFailed(err error) bool {
+	return errgo.Cause(err) == executionFailedError
+}
 
 var invalidConfigError = errgo.New("invalid config")
